pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modern-graphics-generator"
version = "0.1.0"
description = "Generate modern, professional graphics programmatically for articles and presentations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Greg Meyer", email = "greg@gregmeyer.com"}
]
keywords = ["graphics", "visualization", "presentation", "svg", "html", "css"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "playwright>=1.40.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
ai = [
    "openai>=1.0.0",
    "braintrust>=0.0.0",
]

[project.scripts]
modern-graphics = "modern_graphics.cli:main"

[tool.setuptools]
packages = ["modern_graphics"]

[tool.setuptools.package-data]
modern_graphics = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
